SmartCarRemoteControl
It is very good to control you arduino bluetooth car with this app. And
you can control it via the Internet if you  make a little improvement.

Instructions:
First,install the APK on your mobile device.
Second,open your bluetooth car and  search for devices with the app.
Third, connect it.
Fourth,to control the car by send command or press  the buttons.